store: remove last_repair_attempt
