+++ persist/postgres/migrations.go
@@ -449,11 +449,12 @@ FROM objects o;
 	// add consecutive_failed_repairs and next_repair_attempt to slabs
 	func(ctx context.Context, tx *txn, _ *zap.Logger) error {
 		_, err := tx.Exec(ctx, `
+			ALTER TABLE slabs DROP COLUMN last_repair_attempt;
 			ALTER TABLE slabs ADD COLUMN consecutive_failed_repairs SMALLINT NOT NULL DEFAULT 0 CHECK (consecutive_failed_repairs >= 0);
 			ALTER TABLE slabs ADD COLUMN next_repair_attempt TIMESTAMP WITH TIME ZONE NOT NULL DEFAULT NOW();
-			DROP INDEX slabs_id_last_repair_attempt_idx;
+			DROP INDEX slabs_id___idx;
 			DROP INDEX slab_sectors_sector_id_idx;
-			CREATE INDEX slabs_id_last_repair_attempt_idx ON slabs(last_repair_attempt ASC, consecutive_failed_repairs ASC) INCLUDE (next_repair_attempt);
+			CREATE INDEX slabs_id_next_repair_attempt_idx ON slabs(next_repair_attempt ASC);
 			CREATE INDEX contracts_bad_or_inactive_contract_id_idx ON contracts (contract_id) WHERE (NOT good) OR (state NOT IN (0,1)); 
 		`)
 		return err
